chatApp_project/project/services/identity-service/requirements.txt
requests
Flask
firebase-admin
pyjwt
sendgrid
pyyaml